+++ rabbitmq-scenario/src/main/java/org/apache/skywalking/testcase/rabbitmq/Application.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.apache.skywalking.testcase.activemq;
+package org.apache.skywalking.testcase.rabbitmq;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
@@ -25,7 +25,7 @@ import org.springframework.context.annotation.ComponentScan;
 
 @EnableAutoConfiguration
 @SpringBootApplication
-@ComponentScan("org.apache.skywalking.testcase.activemq")
+@ComponentScan("org.apache.skywalking.testcase.rabbitmq")
 public class Application {
 
     public static void main(String[] args) {
